requirements.txt
STreamlit
Pandas
numpy
scikit-learn
mysql-coonector-python
matplotlib
seaborn